frontend/requirements.txt
fastapi==0.115.0
uvicorn==0.23.2
python-dotenv==1.0.1
google-generativeai==0.3.2
